services/local-pod/pyproject.toml
[project]
name = "podex-local-pod"
version = "0.1.0"
description = "Podex Local Pod - Self-hosted compute agent for Podex"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Podex", email = "support@podex.dev" }]
keywords = ["podex", "local-pod", "compute", "docker", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.1.0",
    "python-socketio[asyncio_client]>=5.10.0",
    "docker>=7.0.0",
    "psutil>=5.9.0",
    "structlog>=24.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
podex-local-pod = "podex_local_pod.main:cli"

[project.urls]
Homepage = "https://podex.dev"
Documentation = "https://docs.podex.dev/local-pods"
Repository = "https://github.com/podex/podex"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podex_local_pod"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "docker.*",
    "psutil",
    "psutil.*",
    "structlog.*",
    "socketio.*",
    "aiofiles",
    "aiofiles.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["podex_local_pod.client"]
disallow_untyped_decorators = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
